+++ src/navigation/SettingsStack.tsx
@@ -12,7 +12,7 @@ import About from "../Settings/About"
 import Legal from "../Settings/Legal"
 import ReportIssue from "../Settings/ReportIssue"
 import ENDebugMenu from "../Settings/ENDebugMenu"
-import CallbackFormScreen from "../More/CallbackForm"
+import CallbackFormScreen from "../Settings/CallbackForm"
 import ENLocalDiagnosisKey from "../Settings/ENLocalDiagnosisKeyScreen"
 import ExposureListDebug from "../Settings/ExposureListDebugScreen"
 import { SettingsScreens, SettingsScreen } from "./index"
